Move active_request.py to plugin/core
